render add-employee-panel with JS
