+++ src/renderApp.js
@@ -1,7 +1,7 @@
 import { createCanvasTemplate } from './components/canvas';
 import { createNotificationListTemplate} from './components/notification';
 import { createTemplateForOpenEmployeeListButton, createTemplateForEmployeesListPanel } from './components/employees-list-panel';
-import { createAddEmployeePanelTemplate, createEmployeeAddFormTemplate, createTracksList, createZonesList } from './components/add-employee';
+import { createAddEmployeePanelTemplate, createEmployeeAddFormTemplate, createTracksList, createZonesList } from './components/add-employee-panel';
 
 export const renderApp = () => {
   const rootContainer = document.querySelector('.js-work-display');
@@ -16,13 +16,10 @@ export const renderApp = () => {
   render(rootContainer, createTemplateForEmployeesListPanel(employeesList), 'beforeend');
 
   const employeePanel = rootContainer.querySelector('.js-employee-information-panel');
-  /*const tracks = [{ id: 1, points: [{}, {}]}, { id: 2, points: [{}, {}]}, { id: 4, points: [{}, {}]}];
-  const tracksList = tracks.map(e => createTracksList(e)).join('');
+  const freeTracks = [{ id: 1, points: [{}, {}]}, { id: 2, points: [{}, {}]}, { id: 4, points: [{}, {}]}];
   const zones = [ { id: 1, name: 'Цех 1' }, { id: 2, name: 'Высотные работы' }, { id: 3, name: 'Цех 2' }, { id: 4, name: 'Напряжение' } ];
-  const zonesList = zones.map(e => createZonesList(e)).join('');
-  const form = createEmployeeAddFormTemplate(tracksList, zonesList);
 
-  render(employeePanel, createAddEmployeePanelTemplate(form), 'beforeend');*/
+  render(employeePanel, createAddEmployeePanelTemplate(freeTracks, zones), 'beforeend');
 }
 
 const render = (container, template, position) => {
